Merged assembler and VM; temporarily removed `cadd` instruction.
Corrected README on proper way to install crate.
Removed `cadd` because realistically, implementing it would require an exponential amount of hardware for every bit, making it too expensive. An alternative with "garbage" output is preferred.
Merged assembler and VM because keeping 2 repositories for 1 architecture is too cumbersome when trying to keep both up-to-date.
Simplified processing of arguments and have removed dependency on getopts crate (at least for now).