Instruction fetching method changed; (somewhat) neater printing of register values
Instead of directly copying an instruction in memory into the Instruction Register, the current instruction in the IR is swapped with the next instruction.

That means machine code will be shifting around in memory as the program runs. I have not come up with a way of avoiding this while still being
reversible, but I'm curious what kinds of things will lead from it!