Simplify branches
Turns out you don't entirely need a branch's/assertion's complement in order
to have an `if-else` statement. Instead, you can have the equivalent of 2 `if`
statements, one checking for, say, odd, flipping the parity bit, then having
a second `if` statement checking for odd again which would actually only run for
even numbers (since the parity bit was flipped unconditionally). After the
second branch, the parity bit would have to be flipped again like in the first
one, to ensure the number returns to its original value after the process is
done.

Also replaced on a bunch of zeros with bit shifts for instruction encoding.