The reversible VM is now more reversible!
Ensured that instructions with read- and modify-only registers aren't irreversible.
This is done by swapping the value of the modify-only register(s) with a value of
zero before doing the instruction, so that if the same register is used in a read
position and in a modify position, the value used in the read position is zero.